Move make_file_name to datumaro.util.os_util
It's useful in more than just CLI.

Patch shamelessly stolen from Maxim Zhiltsov's WIP versioning PR. :-)
